Remove circular imports between components/editable/utils and util/utils
